refactors tests and styling
