+++ src/components/pageHeader.styles.js
@@ -1,0 +1,11 @@
+import palette from '../theme/palette.theme'
+
+const styles = {
+  container: {
+    borderBottom: `3px solid ${palette.primary.semi}`,
+    padding: '2rem 0',
+    width: '100%',
+  },
+}
+
+export default styles
